requirements.txt
asgiref==3.7.2
Django==5.0.1
dnspython==2.5.0
sqlparse==0.4.4
typing_extensions==4.9.0
